Add nodejs functions to serverless framework (#15)
