+++ .gitignore
@@ -1,2 +1,6 @@
 .DS_Store
 node_modules
+jspm_packages
+
+# Serverless directories
+.serverless
