+++ circle.yml
@@ -26,6 +26,7 @@ deployment:
     commands:
       - bash install-dependencies.sh
       - apex deploy
+      - npm run deploy
 
 test:
   override:
