--- functions/long-running-success/index.js
@@ -1,9 +1,0 @@
-var iopipe = require('iopipe')({ clientId: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6ImF1dGgwfDU4MTA4ZWU4MDA0NDE0YWU2NzkxMzMyZCIsInVzZXJuYW1lIjoiaW9waXBlX2RlbW8iLCJpYXQiOjE0Nzc0ODAxOTcsImF1ZCI6Imh0dHBzOi8vbWV0cmljcy1hcGkuaW9waXBlLmNvbS9ldmVudC8ifQ.rqy-hDI5x_nSJaQiVUviX5YH6OhzR7HMEQG79d_OuRw" })
-exports.handle = iopipe(
-  function (event, context) {
-    var duration = Math.floor(Math.random()*(2001-900+1)+900);
-    setTimeout(function(){
-      context.succeed('Succeeded with a set duration of ' + duration);
-    }, duration);
-  }
-)
