--- functions/long-running-success/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "iopipe": "^0.2.0"
-  }
-}
