--- functions/random-success-error/index.js
@@ -1,7 +1,0 @@
-var iopipe = require('iopipe')({ clientId: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6ImF1dGgwfDU4MTA4ZWU4MDA0NDE0YWU2NzkxMzMyZCIsInVzZXJuYW1lIjoiaW9waXBlX2RlbW8iLCJpYXQiOjE0Nzc0ODAxOTcsImF1ZCI6Imh0dHBzOi8vbWV0cmljcy1hcGkuaW9waXBlLmNvbS9ldmVudC8ifQ.rqy-hDI5x_nSJaQiVUviX5YH6OhzR7HMEQG79d_OuRw" })
-exports.handle = iopipe(
-  function (event, context) {
-    const shouldError = Math.random() < 0.5;
-    return shouldError ? context.fail('I decided to fail this time.') : context.succeed('I decided to succeed this time.');
-  }
-)
