--- functions/thrown-failure/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "thrown-failure",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "iopipe": "^0.2.1"
-  }
-}
