+++ nodejs/package.json
@@ -1,12 +1,11 @@
 {
-  "name": "random-success-error",
+  "name": "qa-events",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "handler.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
